requirements.txt
pandas~=2.0.3
pyvis~=0.3.2
networkx~=3.1
PyYAML~=6.0.1
openpyxl~=3.1.2
requests~=2.31.0